+++ example/lib/src/main/view/screens/counter/model_counter.dart
@@ -1,0 +1,2 @@
+// Counter app data source
+export 'model/data_source.dart';